Beginning new branch to convert pipeline to floating point.  So far, all Lab processing is done in floating point, and conversion from RGB to Lab is done via LUT and linear interpolation.

+++ rtengine/ipsharpen.cc
@@ -35,7 +35,7 @@ namespace rtengine {
 #define CLIP(a) ((a)>0?((a)<CMAXVAL?(a):CMAXVAL):0)
 #define ABS(a) ((a)<0?-(a):(a))
 
-void ImProcFunctions::dcdamping (float** aI, unsigned short** aO, float damping, int W, int H) {
+void ImProcFunctions::dcdamping (float** aI, float** aO, float damping, int W, int H) {
 
 #ifdef _OPENMP
 #pragma omp for
@@ -55,7 +55,7 @@ void ImProcFunctions::dcdamping (float** aI, unsigned short** aO, float damping,
         }
 }
 
-void ImProcFunctions::deconvsharpening (LabImage* lab, unsigned short** b2) {
+void ImProcFunctions::deconvsharpening (LabImage* lab, float** b2) {
 
     if (params->sharpening.enabled==false || params->sharpening.deconvamount<1)
         return;
@@ -122,7 +122,7 @@ void ImProcFunctions::deconvsharpening (LabImage* lab, unsigned short** b2) {
     delete [] tmpI;
 }
 
-void ImProcFunctions::sharpening (LabImage* lab, unsigned short** b2) {
+void ImProcFunctions::sharpening (LabImage* lab, float** b2) {
 
     if (params->sharpening.method=="rld") {
         deconvsharpening (lab, b2);
@@ -137,26 +137,26 @@ void ImProcFunctions::sharpening (LabImage* lab, unsigned short** b2) {
 #pragma omp parallel
 #endif
     {
-    unsigned short** b3;
+    float** b3;
 
     AlignedBuffer<double>* buffer = new AlignedBuffer<double> (MAX(W,H));
     if (params->sharpening.edgesonly==false) {
 
-        gaussHorizontal<unsigned short> (lab->L, b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
-        gaussVertical<unsigned short>   (b2,     b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
+        gaussHorizontal<float> (lab->L, b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
+        gaussVertical<float>   (b2,     b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
     }
     else {
-        b3 = new unsigned short*[H];
+        b3 = new float*[H];
         for (int i=0; i<H; i++)
-            b3[i] = new unsigned short[W];
+            b3[i] = new float[W];
 
-		bilateral<unsigned short, unsigned int> (lab->L, (unsigned short**)b3, b2, W, H, params->sharpening.edges_radius / scale, params->sharpening.edges_tolerance, multiThread);
-		gaussHorizontal<unsigned short> (b3, b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
-		gaussVertical<unsigned short>   (b2, b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
+		bilateral<float, float> (lab->L, (float**)b3, b2, W, H, params->sharpening.edges_radius / scale, params->sharpening.edges_tolerance, multiThread);
+		gaussHorizontal<float> (b3, b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
+		gaussVertical<float>   (b2, b2, buffer, W, H, params->sharpening.radius / scale, multiThread);
     }
     delete buffer;
 
-    unsigned short** base = lab->L;
+    float** base = lab->L;
     if (params->sharpening.edgesonly)
         base = b3;
 
@@ -182,10 +182,10 @@ void ImProcFunctions::sharpening (LabImage* lab, unsigned short** b2) {
     }
 }
 
-void ImProcFunctions::sharpenHaloCtrl (LabImage* lab, unsigned short** blurmap, unsigned short** base, int W, int H) {
+void ImProcFunctions::sharpenHaloCtrl (LabImage* lab, float** blurmap, float** base, int W, int H) {
 
     int scale = 100 * (100-params->sharpening.halocontrol_amount);
-    unsigned short** nL = base;
+    float** nL = base;
 	#pragma omp parallel for if (multiThread)
     for (int i=2; i<H-2; i++) {
         int max1 = 0, max2 = 0, min1 = 0, min2 = 0, maxn, minn, np1, np2, np3, min, max;
